pyproject.toml
[project]
name = "mesmer"
version = "0.1.0"
description = "Cognitive hacking toolkit for LLMs — treat AI as minds to hack, not software to fuzz"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Galih Laras Prakoso" }
]
keywords = ["llm", "security", "red-team", "cognitive-science", "ai-safety", "jailbreak"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "litellm>=1.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
web = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
]

[project.scripts]
mesmer = "mesmer.interfaces.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100
